packages/google-drive-mcp/pyproject.toml
[project]
name = "mfa-google-drive-mcp"
version = "0.1.0"
description = "MCP server wrapping Google Drive API for agent file management"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "google-api-python-client>=2.0",
    "google-auth>=2.0",
    "pydantic>=2.0",
]

[project.scripts]
google-drive-mcp = "google_drive_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/google_drive_mcp"]
